+++ .changeset/9b8960e8/changes.json
@@ -1,0 +1,38 @@
+{
+  "releases": [{ "name": "@keystone-alpha/session", "type": "major" }],
+  "dependents": [
+    {
+      "name": "@keystone-alpha/admin-ui",
+      "type": "patch",
+      "dependencies": ["@keystone-alpha/session"]
+    },
+    {
+      "name": "@keystone-alpha/server",
+      "type": "patch",
+      "dependencies": ["@keystone-alpha/session"]
+    },
+    {
+      "name": "@keystone-alpha/cypress-project-login",
+      "type": "patch",
+      "dependencies": [
+        "@keystone-alpha/admin-ui",
+        "@keystone-alpha/server",
+        "@keystone-alpha/session"
+      ]
+    },
+    {
+      "name": "@keystone-alpha/cypress-project-twitter-login",
+      "type": "patch",
+      "dependencies": [
+        "@keystone-alpha/admin-ui",
+        "@keystone-alpha/server",
+        "@keystone-alpha/session"
+      ]
+    },
+    {
+      "name": "@keystone-alpha/api-tests",
+      "type": "patch",
+      "dependencies": ["@keystone-alpha/server", "@keystone-alpha/session"]
+    }
+  ]
+}
