+++ .changeset/9b8960e8/changes.md
@@ -1,0 +1,1 @@
+- Create a new package for all session-handling functionality
