+++ .changeset/bc995364/changes.md
@@ -1,0 +1,1 @@
+- Remove the .sessionManager property from the Keystone class
