+++ .changeset/ea436d82/changes.json
@@ -1,0 +1,10 @@
+{
+  "releases": [
+    { "name": "@keystone-alpha/cypress-project-login", "type": "patch" },
+    { "name": "@keystone-alpha/cypress-project-twitter-login", "type": "patch" },
+    { "name": "@keystone-alpha/api-tests", "type": "patch" },
+    { "name": "@keystone-alpha/server", "type": "patch" },
+    { "name": "@keystone-alpha/admin-ui", "type": "patch" }
+  ],
+  "dependents": []
+}
