+++ .changeset/ea436d82/changes.md
@@ -1,0 +1,1 @@
+- Use the new @keystone-alpha/session package
