+++ api-tests/auth-header.test.js
@@ -7,6 +7,7 @@ const bodyParser = require('body-parser');
 const cookieSignature = require('cookie-signature');
 const { multiAdapterRunners } = require('@keystone-alpha/test-utils');
 const { MongooseAdapter } = require('@keystone-alpha/adapter-mongoose');
+const { startAuthedSession, endAuthedSession } = require('@keystone-alpha/session');
 const cuid = require('cuid');
 
 const initialData = {
@@ -60,7 +61,7 @@ function setupKeystone() {
     bodyParser.urlencoded({ extended: true }),
     async (req, res, next) => {
       // Cleanup any previous session
-      await keystone.sessionManager.endAuthedSession(req);
+      await endAuthedSession(req);
 
       try {
         const result = await keystone.auth.User.password.validate({
@@ -72,7 +73,7 @@ function setupKeystone() {
             success: false,
           });
         }
-        await keystone.sessionManager.startAuthedSession(req, result);
+        await startAuthedSession(req, result);
         res.json({
           success: true,
           token: req.sessionID,
