+++ packages/admin-ui/package.json
@@ -43,6 +43,7 @@
     "@emotion/styled": "^10.0.9",
     "@keystone-alpha/field-views-loader": "^1.0.1",
     "@keystone-alpha/fields": "^3.0.1",
+    "@keystone-alpha/session": "^0.0.0",
     "@keystone-alpha/utils": "^2.0.0",
     "apollo-cache-inmemory": "^1.5.1",
     "apollo-client": "^2.5.1",
@@ -52,7 +53,6 @@
     "apollo-upload-client": "^10.0.0",
     "babel-loader": "^8.0.5",
     "babel-plugin-emotion": "^10.0.9",
-    "body-parser": "^1.18.2",
     "chalk": "^2.4.2",
     "compression": "^1.7.4",
     "cross-fetch": "^2.2.0",
