+++ packages/server/WebServer/index.js
@@ -2,13 +2,9 @@ const express = require('express');
 const cors = require('cors');
 const path = require('path');
 const falsey = require('falsey');
-const cookie = require('cookie');
-const expressSession = require('express-session');
-const cookieSignature = require('cookie-signature');
+const { commonSessionMiddleware } = require('@keystone-alpha/session');
 const createGraphQLMiddleware = require('./graphql');
 const initConfig = require('./initConfig');
-
-const COOKIE_NAME = 'keystone.sid';
 
 module.exports = class WebServer {
   constructor(keystone, config) {
@@ -28,58 +24,7 @@ module.exports = class WebServer {
     }
 
     if (Object.keys(keystone.auth).length > 0) {
-      // We have at least one auth strategy
-      // Setup the session as the very first thing.
-      // The way express works, the `req.session` (and, really, anything added
-      // to `req`) will be available to all sub `express()` instances.
-      // This way, we have one global setting for authentication / sessions that
-      // all routes on the server can utilize.
-      function injectAuthCookieMiddleware(req, res, next) {
-        if (!req.headers) {
-          return next();
-        }
-
-        const authHeader = req.headers.authorization || req.headers.Authorization;
-
-        if (!authHeader) {
-          return next();
-        }
-
-        const [type, token] = req.headers['authorization'].split(' ');
-
-        if (type !== 'Bearer') {
-          // TODO: Use logger
-          console.warn(`Got Authorization header of type ${type}, but expected Bearer`);
-          return next();
-        }
-
-        // Split the cookies out
-        const cookies = cookie.parse(req.headers.cookie || '');
-
-        // Construct a "fake" session cookie based on the authorization token
-        cookies[COOKIE_NAME] = `s:${cookieSignature.sign(token, cookieSecret)}`;
-
-        // Then reset the cookies so the session middleware can read it.
-        req.headers.cookie = Object.entries(cookies)
-          .map(([name, value]) => `${name}=${value}`)
-          .join('; ');
-
-        // Always call next
-        next();
-      }
-
-      const sessionMiddleware = expressSession({
-        secret: cookieSecret,
-        resave: false,
-        saveUninitialized: false,
-        name: COOKIE_NAME,
-        store: sessionStore,
-      });
-
-      this.app.use(injectAuthCookieMiddleware, sessionMiddleware);
-
-      // Attach the user to the request for all following route handlers
-      this.app.use(this.keystone.sessionManager.populateAuthedItemMiddleware);
+      this.app.use(commonSessionMiddleware(keystone, cookieSecret, sessionStore));
     }
 
     if (adminUI && adminUI.authStrategy) {
