+++ packages/server/package.json
@@ -9,19 +9,17 @@
   },
   "dependencies": {
     "@keystone-alpha/logger": "^1.0.2",
+    "@keystone-alpha/session": "^0.0.0",
     "@keystone-alpha/utils": "^2.0.0",
     "apollo-errors": "^1.9.0",
     "apollo-server-express": "^2.4.8",
     "body-parser": "^1.18.2",
     "chalk": "^2.4.2",
-    "cookie": "^0.3.1",
-    "cookie-signature": "^1.1.0",
     "cors": "^2.8.4",
     "cuid": "^2.1.6",
     "ensure-error": "^1.0.0",
     "express": "^4.16.3",
     "express-pino-logger": "^4.0.0",
-    "express-session": "^1.15.6",
     "falsey": "^1.0.0",
     "graphql": "^14.0.2",
     "graphql-playground-html": "^1.6.12",
