+++ packages/session/.npmignore
@@ -1,0 +1,2 @@
+**/*.md
+**/*.test.js
