+++ packages/session/README.md
@@ -1,0 +1,12 @@
+---
+section: packages
+title: Session
+---
+
+# Session
+
+This package contains functions to assist with setting up session management in your Keystone system.
+
+```DOCS_TODO
+TODO
+```
