+++ test-projects/login/package.json
@@ -24,6 +24,7 @@
     "@keystone-alpha/fields": "^3.0.1",
     "@keystone-alpha/keystone": "^1.0.4",
     "@keystone-alpha/server": "^2.0.2",
+    "@keystone-alpha/session": "^0.0.0",
     "body-parser": "^1.18.2"
   },
   "devDependencies": {
